refactor(utils): Modularize record existence check functions
Abstracted the record existence check functions from user.service.js and like.service.js into a dedicated module named existanceCheck.util.js.
This enhances code modularity and reusability.
