+++ src/services/user.service.js
@@ -1,25 +1,16 @@
 const { StatusCodes } = require('http-status-codes');
 const { HttpError } = require('../middlewares/errorHandler.middleware');
 const { encryptPassword, issueToken } = require('../utils/authentication.util');
-
-/**
- * @returns {Promise<boolean>} A promise that resolves to a boolean indicating whether the record exists (true) or not (false).
- */
-const isExist = async (conn, email) => {
-    const sql = "SELECT EXISTS (SELECT 1 FROM users WHERE email = ?) AS exist";
-    const [result] = await conn.query(sql, email);
-
-    return result[0].exist === 1;
-};
+const { isUserExist } = require('../utils/existanceCheck.util');
 
 const signUp = async (conn, email, hashedPassword, salt) => {
-    const exists = await isExist(conn, email);
+    const exists = await isUserExist(conn, email);
 
     if (!exists) {
         const sql = "INSERT INTO users (email, password, salt) VALUES (?, ?, ?)";
         const values = [email, hashedPassword, salt];
         const [result] = await conn.query(sql, values);
-    
+
         return result;
     } else {
         throw new HttpError(StatusCodes.CONFLICT, `${email}은 이미 가입된 계정입니다.`)
@@ -41,7 +32,7 @@ const logIn = async (conn, email, password) => {
 };
 
 const passwordResetRequest = async (conn, email) => {
-    const exists = await isExist(conn, email);
+    const exists = await isUserExist(conn, email);
 
     if (exists) {
         return { email };
